+++ java/code/src/com/redhat/rhn/frontend/action/configuration/test/DeleteChannelActionTest.java
@@ -28,7 +28,7 @@ public class DeleteChannelActionTest extends RhnMockStrutsTestCase {
         ConfigChannel cc = ConfigTestUtils.createConfigChannel(user.getOrg());
         ConfigFile cf = ConfigTestUtils.createConfigFile(cc);
 
-        long ccid = cc.getId().longValue();
+        long ccid = cc.getId();
         setRequestPathInfo("/configuration/DeleteChannel");
         addRequestParameter("ccid", "" + ccid);
         actionPerform();
