+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/system/config/test/ServerConfigHandlerTest.java
@@ -160,7 +160,7 @@ public class ServerConfigHandlerTest extends BaseHandlerTestCase {
             if (ActionFactory.TYPE_CONFIGFILES_DEPLOY.getName().
                     equals(action.getTypeName())) {
                 ca = (ConfigAction)ActionManager.lookupAction(regular,
-                                                    action.getId().longValue());
+                        action.getId());
             }
         }
         assertNotNull(ca);
